Make separate StaticAPI module
