+++ server/exe/Main.hs
@@ -1,13 +1,6 @@
-{-# LANGUAGE CPP #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeOperators #-}
-
-#ifdef __SERVE_STATIC__
-{-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE TemplateHaskell #-}
-#endif
 
 import Control.Concurrent.MVar (MVar, modifyMVar, modifyMVar_, newMVar)
 import Control.Monad.IO.Class (liftIO)
@@ -20,15 +13,11 @@ import Servant
 import Servant.API.WebSocket (WebSocket)
 import System.Environment (lookupEnv)
 
-#ifdef __SERVE_STATIC__
-import Data.ByteString (ByteString)
-import qualified Data.ByteString.Lazy as ByteStringL
-import Data.FileEmbed (embedFile)
-#endif
-
 import CategoriesWithFriends (ActiveGame, initGameWithHost, servePlayer)
 import CategoriesWithFriends.Game.Player (PlayerName)
 import CategoriesWithFriends.Logging (debugT)
+
+import StaticAPI (StaticAPI, serverStaticAPI)
 
 type API =
        "game" :> Capture "gameId" Text :> Capture "playerId" PlayerName :> WebSocket
@@ -77,39 +66,3 @@ cleanupGame :: MVar Platform -> Text -> IO ()
 cleanupGame platformVar gameId = do
   modifyMVar_ platformVar $ pure . Map.delete gameId
   debugT $ "Game " ++ show gameId ++ " cleaned up"
-
-{- Serving static files -}
-
-#ifdef __SERVE_STATIC__
-data HTML
-
-instance Accept HTML where
-  contentType _ = "text/html"
-
-instance MimeRender HTML ByteString where
-  mimeRender _ = ByteStringL.fromStrict
-
-type StaticAPI =
-       "static" :> Raw
-  :<|> "game" :> Capture "gameId" Text :> Capture "playerId" PlayerName :> Get '[HTML] ByteString
-  :<|> "game" :> Capture "gameId" Text :> Get '[HTML] ByteString
-  :<|> Get '[HTML] ByteString
-
-serverStaticAPI :: Server StaticAPI
-serverStaticAPI =
-       serveStatic
-  :<|> (\_ _ -> serveHTML)
-  :<|> (\_ -> serveHTML)
-  :<|> serveHTML
-
-serveHTML :: Handler ByteString
-serveHTML = pure $(embedFile "../public/index.html")
-
-serveStatic :: Server Raw
-serveStatic = serveDirectoryFileServer "./public"
-#else
-type StaticAPI = EmptyAPI
-
-serverStaticAPI :: Server StaticAPI
-serverStaticAPI = emptyServer
-#endif
